removed "Embedded resources" section
This section really only existed for CoverageJSON. The methods used were also not that satisfactory and too complicated. In CoverageJSON this is now solved by simply using the "standalone" profile as part of the media type. This also behaves nicely with cross-domain requests (compared to using the Prefer header), see #12. The core issues this spec should tackle are filtering/subsetting and other basics that are relevant for nearly all coverage formats. Therefore, this very specific section is gone now.